% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdsbuilder_options.R
\name{pdsbuilder_options}
\alias{pdsbuilder_options}
\title{Set/get global pdsbuilder options}
\usage{
pdsbuilder_options(...)
}
\arguments{
\item{...}{One of four: (1) nothing, then returns all options as a list; (2) a name of an option element, then returns its value; (3) a name-value pair which sets the corresponding option to the new value (and returns nothing), (4) a list with option-value pairs which sets all the corresponding arguments.}
}
\value{
a list of options, values of an option, or nothing
}
\description{
Global afex options are used, for example, to set the default separator for cell names.
}
\examples{

pdsbuilder_options() # see all options

pdsbuilder_options("sep") # see value of pdsbuilder.sep
\dontrun{
# changes cell separator (e.g., A1.B2)
pdsbuilder_options(sep = ".")
}
}
